+++ parametrization_clean/domain/adaptation/factory.py
@@ -17,8 +17,12 @@ class AdaptationFactory:
     Classes that implement IAdaptationStrategy can be registered and utilized through this factory's registry.
     """
 
-    # Internal registry for available crossover methods
     REGISTRY = {}
+    """Internal registry for available adaptation methods. Users can specify from one of the `algorithm_name` strings
+    available in the dictionary, mapping `algorithm_name` to the corresponding class implementing that algorithm.
+    For example, "xiao" maps to Xiao's adaptation algorithm; users can specify the `adaptation_strategy` in the
+    user config.json file to use this algorithm.
+    """
 
     @classmethod
     def register(cls, algorithm_name: str, adaptation_class):